Fix some variables and functions names
